requirements.txt
Flask
flask-mysqldb
razorpay
gunicorn
mysqlclient
python-dotenv
